cleaning imports (#6770)
* cleaning imports

* added type for static models
