+++ libs/invitation/src/lib/+state/invitation.firestore.ts
@@ -1,7 +1,7 @@
 import { PublicOrganization } from "@blockframes/organization/+state/organization.firestore";
 import type firebase from 'firebase';
 import { PublicUser } from "@blockframes/user/+state/user.firestore";
-import { InvitationStatus, InvitationType } from "@blockframes/utils/static-model";
+import type { InvitationStatus, InvitationType } from "@blockframes/utils/static-model";
 export  { InvitationType, InvitationStatus, invitationStatus } from "@blockframes/utils/static-model";
 type Timestamp = firebase.firestore.Timestamp;
