+++ libs/organization/src/lib/organization/+state/organization.firestore.ts
@@ -1,7 +1,7 @@
 import type firebase from 'firebase';
 import { Location, createLocation } from '@blockframes/utils/common-interfaces/utility';
 import { OrgAppAccess, createOrgAppAccess, Module, App, getAllAppsExcept } from '@blockframes/utils/apps';
-import { OrgActivity, OrganizationStatus } from '@blockframes/utils/static-model/types';
+import type { OrgActivity, OrganizationStatus } from '@blockframes/utils/static-model/types';
 import { createStorageFile, StorageFile, StorageVideo } from '@blockframes/media/+state/media.firestore';
 import { DocumentMeta } from '@blockframes/utils/models-meta';
