+++ libs/utils/src/lib/apps.ts
@@ -1,13 +1,11 @@
 /**
  * Apps definition
  */
-import { OrganizationDocument } from "@blockframes/organization/+state/organization.firestore";
-import { Organization } from "@blockframes/organization/+state/organization.model";
+import { OrganizationBase, OrganizationDocument } from "@blockframes/organization/+state/organization.firestore";
 import { StoreStatus } from "./static-model";
 import { EmailJSON } from '@sendgrid/helpers/classes/email-address';
 import { appUrl } from "@env";
-import { MovieDocument } from "@blockframes/movie/+state/movie.firestore";
-import { Movie } from "@blockframes/movie/+state/movie.model";
+import { MovieBase, MovieDocument } from "@blockframes/movie/+state/movie.firestore";
 import type { RouterQuery } from '@datorama/akita-ng-router-store';
 
 export interface AppMailSetting {
@@ -124,7 +122,7 @@ export function getAppName(slug: App, short = false) {
  * getOrgAppAccess(orgA); // ['catalog', 'festival']
  * getOrgAppAccess(orgB); // ['festival']
  */
-export function getOrgAppAccess(org: OrganizationDocument | Organization, first: App = 'festival'): App[] {
+export function getOrgAppAccess(org: OrganizationDocument | OrganizationBase<Date>, first: App = 'festival'): App[] {
   const apps: App[] = [];
   for (const a of app) {
     const hasAccess = modules.some(m => !!org.appAccess[a]?.[m]);
@@ -144,12 +142,12 @@ export function getOrgAppAccess(org: OrganizationDocument | Organization, first:
 }
 
 /** Return an array of the app access of the movie */
-export function getMovieAppAccess(movie: MovieDocument | Movie): App[] {
+export function getMovieAppAccess(movie: MovieDocument | MovieBase<Date>): App[] {
   return app.filter(a => !['crm'].includes(a) && movie.app[a].access);
 }
 
 /** Return true if the movie has the status passed in parameter for at least one application */
-export function checkMovieStatus(movie: MovieDocument| Movie, status: StoreStatus) {
+export function checkMovieStatus(movie: MovieDocument | MovieBase<Date>, status: StoreStatus) {
   return (Object.keys(movie.app).some(a => movie.app[a].status === status))
 }
 
@@ -162,7 +160,7 @@ export function checkMovieStatus(movie: MovieDocument| Movie, status: StoreStatu
  * getOrgModuleAccess(orgA); // ['dashboard', 'marketplace']
  * getOrgModuleAccess(orgB); // ['marketplace']
  */
-export function getOrgModuleAccess(org: OrganizationDocument | Organization, _a?: App): Module[] {
+export function getOrgModuleAccess(org: OrganizationDocument | OrganizationBase<Date>, _a?: App): Module[] {
   const allowedModules = {} as Record<Module, boolean>;
 
   if (_a) {
